requirements.txt
opencv-python
pandas
numpy
pillow
torch
torchvision
facenet-pytorch
scikit-learn
gradio
